fix: 쿠폰 도메인 Clean + Layered Architecture 적용
