+++ src/main/java/kr/hhplus/be/server/domain/coupon/Coupon.java
@@ -1,0 +1,15 @@
+package kr.hhplus.be.server.domain.coupon;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+public record Coupon(
+    long id,
+    String code,
+    int discount,
+    CouponType type,
+    long issueEndDt,
+    int issueCnt,
+    Long productId
+) {
+
+}
